Converts temperatures fields from Float to Int

+++ app/src/main/java/com/apps/albertmartorell/meteomarto/ui/model/CityUIView.kt
@@ -1,19 +1,17 @@
 package com.apps.albertmartorell.meteomarto.ui.model
 
-import java.util.*
-
 data class CityUIView(
     val latitude: Float? = 0F,
     val longitude: Float? = 0F,
     val main: String? = "",
     val description: String? = "",
     val icon: String? = "",
-    val temperature: Float? = 0F,
+    val temperature: Int? = 0,
     val humidity: Float? = 0F,
     val pressure: Float? = 0F,
-    val temperatureMin: Float? = 0F,
-    val temperatureMax: Float? = 0F,
-    val temperatureFeelsLike: Float? = 0F,
+    val temperatureMin: Int? = 0,
+    val temperatureMax: Int? = 0,
+    val temperatureFeelsLike: Int? = 0,
     val visibility: Long? = 0,
     val speed: Float? = 0F,
     val degrees: Float? = 0F,
